+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Changed
+
+- Core: Category.get_hierarchy now ignores any None objects
+
 ### Fixed
 
 - Admin: Fix primary buttons, list buttons, and filter dropdowns to use css variables